spinloop: use tokio instead of clokwerk
This also changes the behaviour of scheduler. It will now tick fast as
possible, until caught up, instead of emitting an error message.

|     1     |     2     |     3     |     4     |     5     |     6     |
| work -----|          delay          | work | work | work -| work -----|

CMK-15137

+++ v2/robotmk/src/config.rs
@@ -50,7 +50,7 @@ pub struct RobotFrameworkConfig {
 pub struct ExecutionConfig {
     pub n_attempts_max: usize,
     pub retry_strategy: RetryStrategy,
-    pub execution_interval_seconds: u32,
+    pub execution_interval_seconds: u64,
     pub timeout: u64,
 }
